+++ Sources/InventoryCore/Protocols/Components/InventoryComponentLink.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+/// Describes an embedded component that belongs to a parent asset.
+/// Describes an embedded component that belongs to a parent asset.
+public protocol InventoryComponentLink: Sendable {
+    var assetID: UUID { get }
+    var quantity: Int { get }
+    var note: String? { get }
+}